KOGITO-5292: Quarkus - PostgreSQL examples using data index, jobs ser… (#1088)
* KOGITO-5292: Quarkus - Add PostgreSQL persistence to basic and timer 'with-consoles' examples.

* Renamed BD at DB, only copy persistence folder when DB = infinispan and moved kafka conf

+++ kogito-quarkus-examples/process-usertasks-quarkus-with-console/docker-compose/startServices.sh
@@ -1,6 +1,19 @@
 #!/bin/sh
 
-echo "Script requires your Kogito Quickstart to be compiled"
+DB="postgresql"
+
+if [ -n "$1" ]; then
+  if [[ "$1" == "postgresql"  || "$1" == "infinispan" ]];
+  then
+    DB="$1"
+  else
+   echo "Usage: By default postgresql environments is started if no argument is provided"
+   echo "     start POSTGRESQL docker-compose running: ./startServices.sh postgresql or just ./startServices.sh "
+   echo "     start INFINISPAN docker-compose running: ./startServices.sh infinispan "
+   exit 1
+  fi
+fi
+echo "Script requires your Kogito Quickstart to be compiled with the right profile: ../mvn clean install -DskipTests -P$DB"
 
 PROJECT_VERSION=$(cd ../ && mvn help:evaluate -Dexpression=project.version -q -DforceStdout)
 
@@ -22,19 +35,22 @@ elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
    echo "DOCKER_GATEWAY_HOST=172.17.0.1" >> ".env"
 fi
 
-PERSISTENCE_FOLDER=./persistence
-KOGITO_EXAMPLE_PERSISTENCE=../target/classes/META-INF/resources/persistence/protobuf
-
-rm -rf $PERSISTENCE_FOLDER
-
-mkdir -p $PERSISTENCE_FOLDER
-
-if [ -d "$KOGITO_EXAMPLE_PERSISTENCE" ]
+if [ "$1" == "infinispan" ];
 then
+  PERSISTENCE_FOLDER=./persistence
+  KOGITO_EXAMPLE_PERSISTENCE=../target/classes/META-INF/resources/persistence/protobuf
+
+  rm -rf $PERSISTENCE_FOLDER
+
+  mkdir -p $PERSISTENCE_FOLDER
+
+  if [ -d "$KOGITO_EXAMPLE_PERSISTENCE" ]
+  then
     cp $KOGITO_EXAMPLE_PERSISTENCE/*.proto $PERSISTENCE_FOLDER/
-else
-    echo "$KOGITO_EXAMPLE_PERSISTENCE does not exist. Have you compiled the project?"
+  else
+    echo "$KOGITO_EXAMPLE_PERSISTENCE does not exist. Have you compiled the project? mvn clean install -DskipTests -P$DB"
     exit 1
+  fi
 fi
 
 SVG_FOLDER=./svg
@@ -51,4 +67,4 @@ else
     exit 1
 fi
 
-docker-compose up
+docker-compose -f docker-compose-$DB.yml up
