+++ kogito-quarkus-examples/process-usertasks-timer-quarkus-with-console/docker-compose/startServices.sh
@@ -17,24 +17,43 @@ echo "Kogito Image version: ${KOGITO_VERSION}"
 echo "KOGITO_VERSION=${KOGITO_VERSION}" > ".env"
 
 if [ "$(uname)" == "Darwin" ]; then
-   echo "DOCKER_GATEWAY_HOST=kubernetes.docker.internal" >> ".env"
+   echo "Docker-compose don't support host-mode in MAC OS, that is used by jobs-service in this example"
+   exit 1
 elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
    echo "DOCKER_GATEWAY_HOST=172.17.0.1" >> ".env"
 fi
 
-PERSISTENCE_FOLDER=./persistence
-KOGITO_EXAMPLE_PERSISTENCE=../target/classes/META-INF/resources/persistence/protobuf
+DB="postgresql"
 
-rm -rf $PERSISTENCE_FOLDER
+if [ -n "$1" ]; then
+  if [[ "$1" == "postgresql" || "$1" == "infinispan" ]];
+  then
+    DB="$1"
+  else
+   echo "Usage: By default postgresql environments is started if no argument is provided"
+   echo "     start POSTGRESQL docker-compose running: ./startServices.sh postgresql or just ./startServices.sh "
+   echo "     start INFINISPAN docker-compose running: ./startServices.sh infinispan "
+   exit 1
+  fi
+fi
+echo "Have you compiled the project before with the right profile: ../mvn clean install -DskipTests -P$DB"
 
-mkdir -p $PERSISTENCE_FOLDER
-
-if [ -d "$KOGITO_EXAMPLE_PERSISTENCE" ]
+if [ "$1" == "infinispan" ];
 then
+  PERSISTENCE_FOLDER=./persistence
+  KOGITO_EXAMPLE_PERSISTENCE=../target/classes/META-INF/resources/persistence/protobuf
+
+  rm -rf $PERSISTENCE_FOLDER
+
+  mkdir -p $PERSISTENCE_FOLDER
+
+  if [ -d "$KOGITO_EXAMPLE_PERSISTENCE" ]
+  then
     cp $KOGITO_EXAMPLE_PERSISTENCE/*.proto $PERSISTENCE_FOLDER/
-else
-    echo "$KOGITO_EXAMPLE_PERSISTENCE does not exist. Have you compiled the project?"
+  else
+    echo "$KOGITO_EXAMPLE_PERSISTENCE does not exist. Have you compiled the project? mvn clean install -DskipTests -P$DB"
     exit 1
+  fi
 fi
 
 SVG_FOLDER=./svg
@@ -51,5 +70,4 @@ else
     exit 1
 fi
 
-
-docker-compose up
+docker-compose -f docker-compose-$DB.yml up
